+++ packages/edition-slices/src/tiles/tile-p/styles/index.js
@@ -27,8 +27,7 @@ const styles = {
     textAlign: "center"
   },
   imageContainer: {
-    borderColor: colours.functional.contrast,
-    borderWidth: 1,
+    borderRadius: 9999,
     overflow: "hidden",
     width: "30%"
   },