+++ packages/icons/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.9.0](https://github.com/newsuk/times-components/compare/@times-components/icons@2.8.1...@times-components/icons@2.9.0) (2019-04-02)
+
+
+### Features
+
+* save star button package ([#1831](https://github.com/newsuk/times-components/issues/1831)) ([67a39b5](https://github.com/newsuk/times-components/commit/67a39b5))
+
+
+
+
+
 ## [2.8.1](https://github.com/newsuk/times-components/compare/@times-components/icons@2.8.0...@times-components/icons@2.8.1) (2019-04-01)
 
 **Note:** Version bump only for package @times-components/icons